Pine Script code for market sentiment indicators
This code takes the top 40 stocks in the spy and then calculates the percentage change between each close price. Then it weights each one of them according to what percentage of the spy they make up. Then they are all added together and divided by 40. After that the percentage change is displayed on a table.