Pine Script Previous Day High Low
This code automatically plots the previous days high and low. This code only works with stocks from the open of every day to the close of every day. It also has commented out callouts for crossing these lines which if added back in will callout when price crosses these previous day high and low lines.